Updated command

+++ interaction/commands/openai/gpt4.js
@@ -1,5 +1,9 @@
 import { SlashCommandBuilder } from "discord.js";
 import {sleep} from "../../../utils/sleep.js";
+import { generateText } from "../../../utils/OpenAi.js";
+import { config } from "dotenv";
+
+config();
 
 export const data = new SlashCommandBuilder()
     .setName("gpt4")
@@ -10,12 +14,17 @@ export const data = new SlashCommandBuilder()
             .setRequired(true));
 
 export async function execute(interaction) {
-    await interaction.deferReply();
-    await sleep(Math.floor(Math.random()*5000));
-    const response = await interaction.editReply("https://imgur.com/NQinKJB");
-    await sleep(2000);
-    const msg = await interaction.followUp("Today is 1st of April \\:)");
-    await sleep(10000);
-    await response.delete();
-    await msg.delete();
+    if (interaction.user.id!=process.env.MyUserID) {
+        await interaction.deferReply();
+        await sleep(Math.floor(Math.random()*5000));
+        const response = await interaction.editReply("https://imgur.com/NQinKJB");
+        await sleep(2000);
+        const msg = await interaction.followUp("Today is 1st of April \\:)");
+        await sleep(10000);
+        await response.delete();
+        await msg.delete();
+    } else {
+        await interaction.deferReply();
+        await interaction.editReply(`${generateText(interaction.option.getString("prompt"))}`);
+    }
 }
